+++ tests/unit/example.spec.ts
@@ -1,7 +1,7 @@
 import { mount } from '@vue/test-utils';
 import Tab1Page from '@/views/Tab1Page.vue';
 
-jest.mock('@/composables/vault-factory');
+jest.mock('@/composables/useVaultFactory');
 
 describe('Tab1Page.vue', () => {
   it('renders tab 1 Tab1Page', () => {
